added the implementation for creating a new account + switched to another db schema + mapping refactoring

+++ src/main/java/com/symw/entity/Account.java
@@ -10,7 +10,7 @@ public class Account {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name="account_number")
-	private int accountNumber;
+	private Long accountNumber;
 	
 	@Column(name = "name")
 	private String name;
@@ -28,11 +28,11 @@ public class Account {
 	@OneToMany(mappedBy="transactionId", fetch = FetchType.EAGER)
 	private Set<Transaction> transactions;
 
-	public int getAccountNumber() {
+	public Long getAccountNumber() {
 		return accountNumber;
 	}
 
-	public void setAccountNumber(int accountNumber) {
+	public void setAccountNumber(Long accountNumber) {
 		this.accountNumber = accountNumber;
 	}
